+++ CHANGELOG.md
@@ -23,6 +23,7 @@
   * [Ecto.Query] Fix field source when used in `json_extract_path`
   * [Ecto.Query] Properly build CTEs at compile time
   * [Ecto.Repo] Fix `insert_all` query parameter count when using value queries alongside `placeholder`
+  * [Ecto.Repo] Raise if combination query is used in a `many` preload
   * [Ecto.Schema] Ignore associations that aren't loaded on insert
 
 ## v3.8.4 (2022-06-04)